server/requirements.txt
asgiref==3.7.2
channels==4.0.0
Django==5.0.3
django-shortcuts==1.6
djangorestframework==3.14.0
install==1.3.5
pip-autoremove==0.10.0
pytz==2024.1
sqlparse==0.4.4
